feat(roadmaps): delete roadmaps directly from the /roadmaps page

+++ src/actions/roadmap/delete-roadmap.ts
@@ -1,15 +1,25 @@
 'use server';
 import { prisma } from '@/utils/prisma';
 import { revalidateTag } from 'next/cache';
+import { getUserFromSession } from '../user/authed/get-user';
 
-export async function deleteRoadmap(roadmapUid: string, userUid: string) {
+export async function deleteRoadmap(roadmapUid: string) {
+  // get the current user
+  const user = await getUserFromSession();
+
+  if (!user) {
+    throw new Error('User not found');
+  }
+
+  const userUid = user.data.user?.id;
+
   const roadmap = await prisma.userRoadmaps.delete({
     where: {
       uid: roadmapUid,
       AND: {
-        userUid,
-      },
-    },
+        userUid
+      }
+    }
   });
 
   revalidateTag('roadmaps');
